+++ src/editgrouppage.cpp
@@ -70,8 +70,6 @@ void EditGroupPage::showGroup(QString key, std::list<cor::Light> groupDevices, s
 void EditGroupPage::updateDevices(const std::list<cor::Light>& checkedDevices, const std::list<cor::Light>& devices) {
     mSimpleGroupWidget->updateDevices(devices, false);
     mSimpleGroupWidget->setCheckedDevices(checkedDevices);
-
-    resize();
 }
 
 void EditGroupPage::resize() {